refactor: divide admin, public and private layers. Correct dto and model classes
