+++ main-server/src/main/java/ru/practicum/category/controller/CategoryAdminController.java
@@ -10,21 +10,21 @@ import ru.practicum.category.service.CategoryServiceImpl;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Min;
-import java.util.List;
 
 @Valid
 @Slf4j
 @RestController
-public class CategoryController {
+@RequestMapping("/admin/categories")
+public class CategoryAdminController {
 
     private final CategoryServiceImpl categoryService;
 
     @Autowired
-    public CategoryController(CategoryServiceImpl categoryService) {
+    public CategoryAdminController(CategoryServiceImpl categoryService) {
         this.categoryService = categoryService;
     }
 
-    @PostMapping(value = "/admin/categories")
+    @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public CategoryResponseDto createCategory(@RequestBody @Valid CategoryRequestDto category) {
         log.info("POST запрос createCategory - category: \n{}", category);
@@ -33,7 +33,7 @@ public class CategoryController {
         return response;
     }
 
-    @DeleteMapping(value = "/admin/categories/{catId}")
+    @DeleteMapping("/{catId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteCategory(@PathVariable(value = "catId") @Min(1) Long catId) {
         log.info("DELETE запрос deleteCategory - catId: \n{}", catId);
@@ -41,29 +41,12 @@ public class CategoryController {
         log.info("DELETE запрос deleteCategory для catId = {} выполнен успешно", catId);
     }
 
-    @PatchMapping(value = "/admin/categories/{catId}")
+    @PatchMapping("/{catId}")
     public CategoryResponseDto updateCategory(@RequestBody @Valid CategoryRequestDto category,
                                               @PathVariable(value = "catId") @Min(1) Long catId) {
         log.info("PATCH запрос updateCategory - catId: \n{}, category: \n{}", catId, category);
         final CategoryResponseDto response = categoryService.updateCategory(catId, category);
         log.info("PATCH ответ updateCategory - response: \n{}", response);
-        return response;
-    }
-
-    @GetMapping("/categories")
-    public List<CategoryResponseDto> getAllCategories(@RequestParam(value = "from", defaultValue = "0") @Min(0) Integer from,
-                                                      @RequestParam(value = "size", defaultValue = "10") @Min(1) Integer size) {
-        log.info("GET запрос getAllCategories - from: \n{}, size: \n{}", from, size);
-        final List<CategoryResponseDto> response = categoryService.getAllCategories(from, size);
-        log.info("GET ответ getAllCategories - response: \n{}", response);
-        return response;
-    }
-
-    @GetMapping("/categories/{catId}")
-    public CategoryResponseDto getCategoryById(@PathVariable(value = "catId") @Min(1) Long catId) {
-        log.info("GET запрос getCategory - catId: \n{}", catId);
-        final CategoryResponseDto response = categoryService.getCategoryById(catId);
-        log.info("GET ответ getCategory - response: \n{}", response);
         return response;
     }
 }
