+++ main-server/src/main/java/ru/practicum/compilation/controller/CompilationAdminController.java
@@ -11,21 +11,21 @@ import ru.practicum.compilation.service.CompilationServiceImpl;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Min;
-import java.util.List;
 
 @RestController
 @Slf4j
 @Valid
-public class CompilationController {
+@RequestMapping("/admin/compilations")
+public class CompilationAdminController {
 
     private final CompilationServiceImpl service;
 
     @Autowired
-    public CompilationController(CompilationServiceImpl service) {
+    public CompilationAdminController(CompilationServiceImpl service) {
         this.service = service;
     }
 
-    @PostMapping("/admin/compilations")
+    @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public CompilationResponseDto createCompilation(@RequestBody @Valid CompilationRequestDto request) {
         log.info("POST запрос createCompilation - request: \n{}", request);
@@ -34,7 +34,7 @@ public class CompilationController {
         return response;
     }
 
-    @DeleteMapping("/admin/compilations/{compId}")
+    @DeleteMapping("/{compId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteCompilation(@PathVariable(name = "compId") @Min(1) Long compId) {
         log.info("DELETE запрос deleteCompilation - compId: \n{}", compId);
@@ -42,30 +42,12 @@ public class CompilationController {
         log.info("Подборка с id: {} успешно удалено", compId);
     }
 
-    @PatchMapping("/admin/compilations/{compId}")
+    @PatchMapping("/{compId}")
     public CompilationResponseDto updateCompilation(@RequestBody @Valid CompilationUpdateRequestDto request,
                                                     @PathVariable(name = "compId") @Min(1) Long compId) {
         log.info("PATCH запрос updateCompilation - compId: \n{}, request: \n{}", compId, request);
         final CompilationResponseDto response = service.updateCompilation(request, compId);
         log.info("PATCH ответ updateCompilation - response: \n{}", response);
-        return response;
-    }
-
-    @GetMapping("/compilations")
-    public List<CompilationResponseDto> getCompsByParams(@RequestParam(name = "pinned", required = false) Boolean pinned,
-                                                         @RequestParam(name = "from", defaultValue = "0") @Min(0) Integer from,
-                                                         @RequestParam(name = "size", defaultValue = "10") @Min(1) Integer size) {
-        log.info("GET запрос getCompsByParams - pinned: \n{}, from: \n{}, size: \n{}", pinned, from, size);
-        final List<CompilationResponseDto> response = service.getCompsPublic(pinned, from, size);
-        log.info("GET ответ getCompsByParams - response: \n{}", response);
-        return response;
-    }
-
-    @GetMapping("/compilations/{compId}")
-    public CompilationResponseDto getCompById(@PathVariable(name = "compId") @Min(1) Long compId) {
-        log.info("GET запрос getCompById - compId: \n{}", compId);
-        final CompilationResponseDto response = service.getCompById(compId);
-        log.info("GET ответ getCompById - response: \n{}", response);
         return response;
     }
 }
