+++ main-server/src/main/java/ru/practicum/event/dto/EventAdminParams.java
@@ -1,9 +1,9 @@
-package ru.practicum.event.model;
+package ru.practicum.event.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.event.enums.Sorts;
+import ru.practicum.event.enums.EventState;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -11,13 +11,15 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class EventPublicParams {
+public class EventAdminParams {
 
-    private String text;
+    private List<Long> users;
+
+    private List<EventState> states;
+
     private List<Long> categories;
-    private Boolean paid;
+
     private LocalDateTime rangeStart;
+
     private LocalDateTime rangeEnd;
-    private Boolean onlyAvailable;
-    private Sorts sort;
 }
