+++ main-server/src/main/java/ru/practicum/event/dto/EventPublicParams.java
@@ -1,9 +1,9 @@
-package ru.practicum.event.model;
+package ru.practicum.event.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.event.enums.EventState;
+import ru.practicum.event.enums.Sorts;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -11,15 +11,13 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class EventAdminParams {
+public class EventPublicParams {
 
-    private List<Long> users;
-
-    private List<EventState> states;
-
+    private String text;
     private List<Long> categories;
-
+    private Boolean paid;
     private LocalDateTime rangeStart;
-
     private LocalDateTime rangeEnd;
+    private Boolean onlyAvailable;
+    private Sorts sort;
 }
