+++ main-server/src/main/java/ru/practicum/event/dto/EventRequestDto.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.event.model.Location;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
