+++ main-server/src/main/java/ru/practicum/event/dto/EventRequestStatusUpdateResult.java
@@ -1,8 +1,9 @@
-package ru.practicum.request.dto;
+package ru.practicum.event.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import ru.practicum.request.dto.ParticipationResponseDto;
 
 import java.util.List;
