+++ main-server/src/main/java/ru/practicum/request/controller/RequestPrivateController.java
@@ -4,8 +4,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.request.dto.EventRequestStatusUpdateRequest;
-import ru.practicum.request.dto.EventRequestStatusUpdateResult;
 import ru.practicum.request.dto.ParticipationResponseDto;
 import ru.practicum.request.service.RequestServiceImpl;
 
@@ -14,16 +12,17 @@ import java.util.List;
 
 @RestController
 @Slf4j
-public class RequestController {
+@RequestMapping("/users/{userId}/requests")
+public class RequestPrivateController {
 
     private final RequestServiceImpl service;
 
     @Autowired
-    public RequestController(RequestServiceImpl service) {
+    public RequestPrivateController(RequestServiceImpl service) {
         this.service = service;
     }
 
-    @GetMapping("/users/{userId}/requests")
+    @GetMapping
     public List<ParticipationResponseDto> getUserRequests(@PathVariable(name = "userId") @Min(1) Long userId) {
         log.info("GET запрос getUserRequests - userId: \n{}", userId);
         final List<ParticipationResponseDto> response = service.getUserRequests(userId);
@@ -31,7 +30,7 @@ public class RequestController {
         return response;
     }
 
-    @PostMapping("/users/{userId}/requests")
+    @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public ParticipationResponseDto createRequest(@PathVariable(name = "userId") @Min(1) Long userId,
                                                   @RequestParam(name = "eventId") @Min(1) Long eventId) {
@@ -41,31 +40,12 @@ public class RequestController {
         return response;
     }
 
-    @PatchMapping("/users/{userId}/requests/{requestId}/cancel")
+    @PatchMapping("/{requestId}/cancel")
     public ParticipationResponseDto cancelRequestByUser(@PathVariable(name = "userId") @Min(1) Long userId,
                                                         @PathVariable(name = "requestId") @Min(1) Long requestId) {
         log.info("PATCH запрос cancelRequestByUser - userId: \n{}, requestId: \n{}", userId, requestId);
         final ParticipationResponseDto response = service.cancelRequestByUser(userId, requestId);
         log.info("PATCH ответ cancelRequestByUser - response: \n{}", response);
-        return response;
-    }
-
-    @GetMapping("/users/{userId}/events/{eventId}/requests")
-    public List<ParticipationResponseDto> getEventRequests(@PathVariable(name = "userId") @Min(1) Long userId,
-                                                           @PathVariable(name = "eventId") @Min(1) Long eventId) {
-        log.info("GET запрос getEventRequests - userId: \n{}, eventId: \n{}", userId, eventId);
-        final List<ParticipationResponseDto> response = service.getEventRequests(userId, eventId);
-        log.info("GET ответ getEventRequests - response: \n{}", response);
-        return response;
-    }
-
-    @PatchMapping("/users/{userId}/events/{eventId}/requests")
-    public EventRequestStatusUpdateResult changeRequestsStatus(@RequestBody EventRequestStatusUpdateRequest request,
-                                                               @PathVariable(name = "userId") @Min(0) Long userId,
-                                                               @PathVariable(name = "eventId") @Min(1) Long eventId) {
-        log.info("PATCH запрос changeRequestsStatus - userId: \n{}, eventId: \n{}, request: \n{}", userId, eventId, request);
-        final EventRequestStatusUpdateResult response = service.changeRequestsStatus(userId, eventId, request);
-        log.info("PATCH ответ changeRequestsStatus - response: \n{}", response);
         return response;
     }
 }
